feat: zerocopy Records (postpone BatchView)
This closes https://github.com/tisonkun/kafka-api/issues/2.

The idiom to consume records is with record batch. Thus, we implement
into_batches and take_batches methods. But it's possible to use a
(mutatble) batch view to read and write records. It's included in the
previous commit and we can restore it if there is any real world usage.

BTW, simplesrv can serve as a simple Kafka server with basic pubsub full
featured now. An improvement point is filtering records on fetch
responses based on fetch_offset. But it's not a bug because the consumer
will filter records by fetch_offset also.

Signed-off-by: tison <wander4096@gmail.com>